+++ templates/course/parent-course.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * Back to Course Template
+ * @since  1.0.0
+ * @version 3.0.0
+ */
+
+if ( ! defined( 'ABSPATH' ) ) { exit; }
+
+global $post;
+
+$lesson = new LLMS_Lesson( $post->ID );
+
+printf( __( '<p class="llms-parent-course-link">Back to: <a class="llms-lesson-link" href="%s">%s</a></p>', 'lifterlms' ), get_permalink( $lesson->get_parent_course() ), get_the_title( $lesson->get_parent_course() ) );